+++ dart/geocore/lib/src/base/bounds.dart
@@ -93,9 +93,15 @@ abstract class Bounds<T extends Point> extends Geometry
     return buf.toString();
   }
 
-  /// Returns new bounds transformed from this bounds using [transform].
+  /// Returns new bounds transformed from this bounds using [transformation].
   @override
-  Bounds<T> transform(TransformPoint transform);
+  Bounds<T> transform(TransformPoint transformation);
+
+  @override
+  Bounds<R> project<R extends Point>(
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  });
 
   /// Returns true if this bounds intesects with [other] bounds in 2D.
   ///
@@ -207,9 +213,19 @@ class BoundsBase<T extends Point> extends Bounds<T> with EquatableMixin {
   }
 
   @override
-  Bounds<T> transform(TransformPoint transform) => BoundsBase(
-        min: min.transform(transform) as T,
-        max: max.transform(transform) as T,
+  Bounds<T> transform(TransformPoint transformation) => BoundsBase(
+        min: min.transform(transformation) as T,
+        max: max.transform(transformation) as T,
+      );
+
+  @override
+  Bounds<R> project<R extends Point>(
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  }) =>
+      BoundsBase(
+        min: min.project(projection, to: to),
+        max: max.project(projection, to: to),
       );
 }
 
@@ -257,8 +273,15 @@ class _LazyBounds<T extends Point> extends Bounds<T> {
       _ensureBounds().newFrom(coords, offset: offset, length: length);
 
   @override
-  Bounds<T> transform(TransformPoint transform) =>
-      _ensureBounds().transform(transform);
+  Bounds<T> transform(TransformPoint transformation) =>
+      _ensureBounds().transform(transformation);
+
+  @override
+  Bounds<R> project<R extends Point>(
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  }) =>
+      _ensureBounds().project(projection, to: to);
 
 /*
   // See lint => avoid_equals_and_hash_code_on_mutable_classes
@@ -294,4 +317,11 @@ class _EmptyBounds extends Bounds {
 
   @override
   Bounds transform(TransformPoint transform) => this;
+
+  @override
+  Bounds<R> project<R extends Point>(
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  }) =>
+      throw const FormatException('Cannot project empty bounds.');
 }