+++ dart/geocore/lib/src/base/geometry.dart
@@ -41,25 +41,22 @@ abstract class Geometry extends Bounded {
   /// True if this geometry is NOT considered empty without data or coordinates.
   bool get isNotEmpty => !isEmpty;
 
-  /// Returns a new geometry with all points transformed using [transform].
+  /// Returns a new geometry with all points transformed using [transformation].
   ///
   /// The transformed geometry object must be of the same geometry type with
   /// this object.
   @override
-  Geometry transform(TransformPoint transform);
+  Geometry transform(TransformPoint transformation);
 
-/* 
-  todo: how to define this right? 
-
-  /// Returns a new geometry with all points projected using [project] function.
+  /// Returns a new geometry with all points projected using [projection].
   ///
-  /// When [factory] is provided, then target points of [R] are created using
-  /// that as a point factory. Otherwise [project] uses it's own factory.
+  /// When [to] is provided, then target points of [R] are created using
+  /// that as a point factory. Otherwise [projection] uses it's own factory.
+  @override
   Geometry project<R extends Point>(
-    ProjectPoint<Point, R> project, [
-    PointFactory<R>? factory,
-  ]);
-*/
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  });
 }
 
 /// An empty (non-existent) geometry as an private implementation.
@@ -83,14 +80,12 @@ class _EmptyGeometry extends Geometry with EquatableMixin {
   @override
   Geometry transform(TransformPoint transform) => this;
 
-/*
   @override
   Geometry project<R extends Point>(
-    ProjectPoint<Point, R> project, [
-    PointFactory<R>? factory,
-  ]) =>
+    ProjectPoint<R> projection, {
+    PointFactory<R>? to,
+  }) =>
       throw const FormatException('Cannot project empty geometry.');
-*/
 
   @override
   List<Object?> get props => [];